// src/globals/models/enums/LabelTypeEnum.ts
export enum LabelTypeEnum {
  ERROR_LABEL = "errorLabel",
  HEADLINE_LABEL = "headlineLabel",
  QUESTION_LABEL = "questionLabel",
  QUIZ_BUTTON_LABEL = "quizButtonLabel",
  SCORE_LABEL = "scoreLabel",
  SCORE_SUB_LINE_LABEL = "scoreSubLineLabel",
  SUB_LINE_LABEL = "subLineLabel",
  SUBMIT_BUTTON_LABEL = "submitButtonLabel",
  TOPIC_LABEL = "topicLabel",
}